Update User_Analytics_Behaviour.json
Made changes to parameter assignment in queries.

Instead of parsing the value in the parameter "Enter User Principal Name" this is now parsed based on the clicked record in the top identity table, where the analyst selects the applicable user for their analysis. The UserPrincipleName (in the top identitiy table) is then being parsed in each query.